Changing datatype to size_t

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -8,7 +8,7 @@
  */
 size_t print_listint_safe(const listint_t *head)
 {
-	unsigned int num_of_nodes = 0;
+	size_t num_of_nodes = 0;
 	long int diff;
 
 	while (head)
